Common Data Format Report version 2.0
These changes were made after talking to
Shreyas Dusane of Highmark

Instead of Comma ( , ) now the Piping
operator ( | ) is used as the separator.

Three separate files are generated for each segment.

The Address file and the Loan/Account file are related to Client file
by parent ID which contains the Clients ID