Support #1140: Payment of fees for all loans
A rake task (rake mostfit:fees:pay) has been
added that shall pay the fees for all the loans
for which payment of fees is not recorded. This
rake task is designed in such a manner such that
it can be used for any case other that this particular
support request.

For this particular support request #1140, first you need to run

$ rake mostfit:fees:levy

and then only

$ rake mostfit:fees:pay

If the above sequence is not followed no fees might be paid at all.