Feature #676: Daily Transaction Summary now give center-wise breakup
The daily transaction summary now gives a center-wise
report when a particular branch is selected. Many of the
LoanHistory functions that were previously used to generate
the report are now no longer working properly, hence
they are not being used in this report. The report is slower
to generate but here the emphasis is on correctness and
not on speed.

Also a major cause of worry is the unreliability of the
LoanHistory model incase of recording the preclosures/
prepayments of Loans. This report is not guaranteed to
give the correct preclosure amount due to this
ambiguity.

Conflicts:

	app/models/reports/daily_transaction_summary.rb