Bug #1690:(Resolved) The insurance policy attached to a loan now gets saved
The datamapper functionality that saves the children objects while
saving the parent objects; in this case insurance policy linked to the
loan getting saved along with the loan is not working.

I have not been able to find out the reason why it is not doing so.

Hence this patch.