The Highmark response is processed and the loans are appropriately flagged
The Highmark response is evaluated and is assigned a status accepted
or rejected based on certain paramaters. Currently
a reject occurs if the client has more then two loans or the
total loan amount on him is greater than equal to Rs. 50,000.
This can later be changed only changes in the code but later
should be made user customizable.

Apart from the that the view of the Highmark responses
has been improved and also the the background color flag
has been rectified.