Feature #1814: Added a new hierarchy called a district
This feature has been added at the request of Suryoday
Microfinance.

This model does NOT break the existing heirarchy of

Regions
  |
Areas
  |
Districts
  |
Branches

but adds to it like

  Regions
    |
 A r e a s
    |   \
    |  Districts
    |   /
  Branches

This commit lacks the AJAX functionality in the view side.
It is not working properly and will be corrected in subsequent commits