Feature #1810: Funding line behaviour changed in bulk loan creation screen
1) Added a field funding line for Mfi which can be configured from
Profile and System wide polices as the default funding line.

2) Also added a boolean in Mfi which configures if the funding line
selection has to be made default or made configurable for bulk loan
creations.

Behaviour:

If the funding line has been defaulted to a particular one and if the
choice of funding line in bulk loan creations is not allowed for the whole
system the bulk loan creation screen does not give an option to change
it.

But if a choice has been allowed of funding lines in bulk loan creations
then a dropdown is given in the bulk loan creation screen