Bug #1948: Integrity error resolved
The error was arising due to an improperly generated SQL query
by the passing of an rogue argument from a FundingLine instance function
to the LoanHistory functions. The argument has now been rectified.

MySQL was not able to understand the funding_line_id paramater belonged
to which table and hence it gave an integrity error