SQL query added to dissociate staff from users with role of data_entry
You will also see a rake task for this. Originally a rake task was
written to solve this problem. The statement in the rake task which
nullifies the staff_member of the user works on the interactive merb but
unfortunately does not work in the rake task.

user.save and user.save! don't give the value true, atleast on my
machine

the only way to do this is via the attached SQL script

so why has the rake task been committed?

The rake task will generate a csv file in docs/rake_logs
folder listing all the users and their corresponding staffmembers that
have been dissociated. The rake task is for purely communication
purposes between the team members. It can be deleted in the subsequent
commits using git rm. Also who knows the rake task might work on your
machine.

If you don't want to use the rake task you can use the following SQL
query to get the list of the users with role data_entry who are
associated with them:-

select staff_members.name, users.login  from users, staff_members where
users.role = 1 and staff_members.user_id = users.id;