Bug #2103: Meetings are now shifted due to creation of Holiday calendars
The bug was in the holidays function of the branch model. It made the
assumption that if a holiday calendar already exists either for a branch
then a holiday calendar would not exist for its corresponding area or
region and hence it did not return the new holidays that were applied to
an area or a region. This function has now correctly returns
all the holidays that maybe applicable to a branch.