Shows the list of clients for a selected center
This gives the list for clients that belong to a center in the bulk loan
applications screen. The clients that are ineligible for loan
applications are grayed out  but not disabled. The form for submitting
the clients does NOT work. The controller code has not yet been written

TODO: to make the display of the clients list AJAXY