Story 26336923: Initial models for recording the overlap_report_requests and overlap_report_responses
A request object is supposed to be created everytime a request row is
added to the overlap_report_request csv file by the rake task. This
functionality will be added in the future.

A response object is supposed to be created for every request summarizing the
responses received per request in the response xml file