Story 26424825: Rate loan application on credit bureau overlap report
1. The credit bureau responds with an overlap report for each request on
new loan applications that are communicated to it

2. The following conditions are to be considered when rating a loan
application on which an overlap report has been received

a. The sum of the (current indebtedness of loan applicant) and (proposed
new loan) must not exceed Rs.50,000/- (to be controlled via
administrative or application configuration)

b. The total number of loans (with the addition of the proposed loan)
should not exceed two (to be controlled via administrative or
application configuration)

3. If any of the above conditions are not met, the loan application is
rated 'negatively', otherwise it is rated 'positively'