center cycle model rewritten
Previously when center cycles were ended the the same center cycle was
updated with a new initiated on date and restarted instead of a new
model of center cycle being created. That error has now been rectified.
Now a new instance of center cycle is created and persisted whenever a
new cycle begins.

Basic validations have been added to the model, like
1. the cycle number of the current center cycle is always 1 more than
   the cycle number of the previous center cycle
2. the current center cycle does not begin before the previous center
   cycle closed on date

more validations have to be added and tests to be written in the
upcoming commits