Encumberances are now displayed alongwith various calculations as per the passed date.
The values displayed are Assigned Value, Portfolio Scheduled Value and Portfolio Actual Value and Excess/Shortfall.
The calculations are done as per the date passed by the user.

Added a link to display details of each encumberance for each page.

Modified the corresponding methods in the facade and the model for the same.
Removed the show action in the controller as it wasn't of any use anymore.